Added missing table + fixed light on one table
There was a table missing on the left hand side (when exiting from player spawn), so I placed that in the scene. Also Emy mentioned that one of the tables had the lights strangely flickering when you would walk up to it, so I seemed to have fixed it by just deleting that one and adding in a new table.