Meters spawn in /w Anteater
KeyPickUp script (which does not do only one thing lmaoo, it's so bloated :skull:) now adds in the awareness and danger UI elements when the anteater spawns in. Before they would be in the level regardless. Also made some cute changes to the Anteater's inspector values that effect the meter.

:dancers: :dancers: :dancers: :dancers: :dancers: :dancers: :dancers: