Fixed Phone Bugs + Updated for "Pre-Loaded" Msgs
The Phone canvas had the first Text Image directly attached to it which meant that when the Phone script ran, a copy of the first image would appear above it. I removed the Text Image of the first message from the Telephone in the In-Game Canvas.

Incorrect time was set for message "ARE" which made it appear later than the string of messages which should say "DUDE WHERE ARE YOU"

We want the first 2 messages to be "pre-loaded" (i.e. they're seemingly sent before the game starts, so there should be no animation or sfx for them). I added a "isPreloaded" bool to the Phone script and changed it so messages marked as such are displayed in Start() but do not play the animation or eerieSfx. Also renamed variable "sound" in Phone.cs to "playEerieSfx"